@@ -4,69 +4,66 @@
   "description": "A skeleton for react based applications",
   "main": "index.js",
   "scripts": {
-    "start": "cross-env NODE_ENV=development node server/index.js",
-    "build": "cross-env NODE_ENV=production webpack --config webpack/webpack.config.production.js --optimize-minimize",
+    "start": "ts-node .webpack/dev-server",
+    "build": "wp --config .webpack/config.production.ts",
+    "serve": "serve ./dist",
     "format": "prettier --write \"src/**/*.ts\"",
     "code:check": "yarn code:lint; yarn code:format --check",
     "code:clean": "yarn code:lint --fix; yarn code:format --write",
     "code:lint": "eslint --ext .js,.jsx,.ts,.tsx \"src/\"",
-    "code:format": "prettier './src/**/**.{js,jsx,ts,tsx}'"
+    "code:format": "prettier './src/**/**.{js,jsx,ts,tsx}'",
+    "prepare": "husky install"
   },
   "author": "olefirenk@gmail.com",
   "license": "MIT",
   "dependencies": {
-    "eslint-plugin-react": "^7.19.0",
-    "prop-types": "^15.7.2",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-is": "^17.0.1"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-is": "^18.2.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.1.2",
-    "@babel/eslint-parser": "^7.17.0",
-    "@babel/plugin-proposal-class-properties": "^7.1.0",
-    "@babel/plugin-proposal-decorators": "^7.1.2",
-    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.8.3",
-    "@babel/plugin-proposal-optional-chaining": "^7.8.3",
-    "@babel/plugin-transform-runtime": "^7.17.0",
-    "@babel/preset-env": "^7.1.0",
-    "@babel/preset-react": "^7.0.0",
-    "@babel/preset-typescript": "^7.1.0",
-    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.4",
-    "@types/react": "^17.0.2",
-    "@types/react-dom": "^17.0.1",
-    "@types/react-router": "^5.1.1",
-    "@types/react-router-config": "^5.0.1",
-    "@types/react-router-dom": "^5.1.0",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.7",
+    "@types/express": "^4.17.13",
+    "@types/node": "^18.0.0",
+    "@types/progress-bar-webpack-plugin": "^2.1.2",
+    "@types/react": "^18.0.14",
+    "@types/react-dom": "^18.0.5",
     "@types/webpack-env": "^1.13.6",
+    "@types/webpack-hot-middleware": "^2.25.6",
     "@typescript-eslint/eslint-plugin": "^5.13.0",
     "@typescript-eslint/parser": "^5.13.0",
-    "babel-loader": "^8.0.4",
-    "babel-plugin-transform-typescript-metadata": "^0.3.2",
-    "cross-env": "^7.0.2",
+    "copy-webpack-plugin": "^11.0.0",
     "eslint": "^8.10.0",
     "eslint-config-prettier": "^8.4.0",
+    "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-react": "^7.19.0",
     "eslint-plugin-react-hooks": "^4.0.0",
     "eslint-webpack-plugin": "^3.1.1",
     "express": "^4.16.3",
     "file-loader": "^6.0.0",
     "fork-ts-checker-webpack-plugin": "^7.2.1",
-    "html-webpack-plugin": "^5.1.0",
-    "husky": "^7.0.4",
-    "lint-staged": "^12.3.4",
-    "prettier": "^2.0.5",
-    "pretty-quick": "^3.1.0",
+    "html-webpack-plugin": "^5.5.0",
+    "husky": ">=6",
+    "lint-staged": ">=10",
+    "prettier": "^2.7.1",
     "progress-bar-webpack-plugin": "^2.1.0",
-    "react-refresh": "^0.11.0",
-    "react-svg-loader": "^3.0.3",
-    "terser-webpack-plugin": "^5.1.1",
+    "react-refresh": "^0.13.0",
+    "react-refresh-typescript": "^2.0.5",
+    "serve": "^13.0.2",
+    "terser-webpack-plugin": "^5.3.3",
+    "ts-loader": "^9.3.1",
+    "ts-node": "^10.8.1",
     "typescript": "^4.1.5",
     "url-loader": "^4.1.0",
-    "webpack": "^5.23.0",
-    "webpack-dev-middleware": "^5.3.1",
+    "webpack": "^5.73.0",
+    "webpack-dev-middleware": "^5.3.3",
     "webpack-hot-middleware": "^2.25.1",
-    "webpack-merge": "^5.7.3",
+    "webpack-merge": "^5.8.0",
     "webpack-nano": "^1.1.1"
+  },
+  "lint-staged": {
+    "*.{js,jsx,ts,tsx}": "eslint --cache --fix",
+    "*.{js,jsx,ts,tsx,json,gql,html,md,yml}": "prettier --write"
   }
 }
